fix: move reusable components to common folder
